Bug fix, gains to 16bits variable, corrected Condition forces, and more

+++ FFBJoystick/Inc/PIDReportType.h
@@ -367,7 +367,7 @@ typedef struct
   int16_t startMagnitude;
   int16_t endMagnitude;
   uint16_t period; // 0..32767 ms
-  uint16_t duration, fadeTime, attackTime, elapsedTime;
+  uint32_t duration, fadeTime, attackTime, elapsedTime;
   uint64_t startTime;
   bool useEnvelope;
 } TEffectState;
